+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 - [docs:Add code coverage blocking.](https://github.com/Tencent/spring-cloud-tencent/pull/837)
 - [docs:Add license checker GitHub Action.](https://github.com/Tencent/spring-cloud-tencent/pull/841)
+- [refactor:move loadbalancer to discovery module.](https://github.com/Tencent/spring-cloud-tencent/pull/845)